KAFKA-14274, #4: introducing FetchBuffer and FetchCollector
More abstraction of the Fetcher code to split into classes to buffer raw results (FetchBuffer) and to collect raw results into ConsumerRecords (FetchCollector).
